+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 Note: This is in reverse chronological order, so newer entries are added to the top.
 
+# v3.1.6
+15 June 2026
+
+### IBM Verify MFA SDK for iOS
+* Increment version
+
+### IBM Verify Core SDK for iOS
+* Increment version
+
+### IBM Verify Authentication SDK for iOS
+* Support for different date formats when decoding `TokenInfo.expiresOn`.
+* Increment version
+
+### IBM Verify Adaptive SDK for iOS
+* Increment version
+
+### IBM Verify FIDO2™ SDK for iOS
+* Increment version
+<br/>
+<br/>
+
 # v3.1.5
 12 June 2026